abstracting functions from components and unit tests
