+++ src/App.test.js
@@ -1,11 +1,10 @@
-import { render, screen } from '@testing-library/react';
+import { render } from '@testing-library/react';
 import App from './App';
 
-import * as React from 'react';
+// import * as React from 'react';
 
 
-// test('renders learn react link', () => {
-//   render(<App />);
-//   const linkElement = screen.getByText(/learn react/i);
-//   expect(linkElement).toBeInTheDocument();
-// });
+test('App test', () => {
+  const linkElement = 1 + 1;
+  expect(linkElement).toBe(2);
+});
