+++ src/components/Introduction/Introduction.js
@@ -29,5 +29,4 @@ const Introduction = () => {
     );
 }
 
-
 export default Introduction;
