+++ src/index.js
@@ -2,7 +2,7 @@ import React from 'react';
 import ReactDOM from 'react-dom/client';
 import { Provider } from 'react-redux';
 
-import store from './store/redux';
+import store from './store/store';
 
 import './index.css';
 import App from './App';
