+++ src/store/store.js
@@ -1,0 +1,24 @@
+import { legacy_createStore } from 'redux';
+
+const initialState = {
+    propertyValue: 350000,
+    
+};
+
+// Reducer
+const handleState = (state = initialState, action) => {
+    switch (action.type) {
+        case 'SET_VALUE':
+        return { 
+            ...state,
+            propertyValue: action.value
+        };
+        default:
+            return state
+    }  
+};
+
+const store = legacy_createStore(handleState);
+
+export default store;
+
